requirements.txt
activecampaign-python==1.0.0
astroid==2.3.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-bigquery==1.23.1
google-cloud-core==1.1.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
urllib3==1.25.7
wrapt==1.11.2
